moved deletion logic to user delete queue service and added child container registration
